removed diffraction subpackage, absorbed its contents into other modules
