+++ xrsdkit/scattering/structure_factors.py
